ci: Remove auto-formatting on push
At first glance, it seemed like a good idea: any code contribution was automatically formatted. So you don't need care about it.

But it is more disturbing. In some cases, this breaks the protection of the branches. Tests can be failed by a commit. But since the last commit is formatting, and tests are not run for automatic commit, GH protection will allow such a merge. Even though the previous commit broke the tests.