test: tests for each package are run sequentially
Previously, tests for each package and for each OS were run in parallel. Now only per-OS tests are run in parallel, and per-package tests are run sequentially.

This does not degrade the overall test execution time, but it makes the workflow files simpler and easier to understand.

It also reduces financial costs, as the total processing time decreases.